refactor(config): reorganize schemas and normalize firewall behavior shape
- move schema files from helpers/ to a dedicated schemas/ directory
and extract the firewall schema into its own module (firewall/behaviors,
firewall/functionsInstances).
- replace the property-per-behavior pattern ({ deny: true },
{ runFunction: ... }, { setWafRuleset: {...} }) with a uniform
{ type, attributes } shape across firewall behaviors, aligning the
config contract with the API representation.
