+++ packages/config/src/configProcessor/schemas/schema.ts
@@ -11,12 +11,6 @@ import {
   EDGE_CONNECTOR_LOAD_BALANCE_METHOD,
   EDGE_CONNECTOR_TRANSPORT_POLICY,
   EDGE_CONNECTOR_TYPES,
-  FIREWALL_RATE_LIMIT_BY,
-  FIREWALL_RATE_LIMIT_TYPES,
-  FIREWALL_RULE_CONDITIONAL,
-  FIREWALL_RULE_OPERATORS,
-  FIREWALL_VARIABLES,
-  FIREWALL_WAF_MODES,
   HEADER_BEHAVIORS,
   ID_BEHAVIORS,
   NETWORK_LIST_TYPES,
@@ -31,6 +25,8 @@ import {
   WORKLOAD_TLS_VERSIONS,
   WORKLOADS_ACCESS_TYPES,
 } from '../../constants';
+
+import firewallSchema from './firewall';
 
 const createCriteriaBaseSchema = (isRequestPhase = false) => ({
   type: 'object',
@@ -483,170 +479,6 @@ const schemaKV = {
   errorMessage: {
     additionalProperties: 'No additional properties are allowed in kv items.',
     required: "The 'name' field is required.",
-  },
-};
-
-const firewallRulesBehaviorsSchema = {
-  type: 'array',
-  minItems: 1,
-  maxItems: 10,
-  items: {
-    type: 'object',
-    oneOf: [
-      {
-        properties: {
-          runFunction: {
-            type: ['string', 'number'],
-            errorMessage: "The 'runFunction' behavior must be a string or number",
-          },
-        },
-        required: ['runFunction'],
-        additionalProperties: false,
-      },
-      {
-        properties: {
-          setWafRuleset: {
-            type: 'object',
-            properties: {
-              wafMode: {
-                type: 'string',
-                enum: FIREWALL_WAF_MODES,
-                errorMessage: `The wafMode must be one of: ${FIREWALL_WAF_MODES.join(', ')}`,
-              },
-              wafId: {
-                type: ['string', 'number'],
-                errorMessage: 'The wafId must be a string or number',
-              },
-            },
-            required: ['wafMode', 'wafId'],
-            additionalProperties: false,
-            errorMessage: {
-              additionalProperties: 'No additional properties are allowed in the setWafRuleset object',
-              required: "Both 'wafMode' and 'wafId' fields are required in setWafRuleset",
-            },
-          },
-        },
-        required: ['setWafRuleset'],
-        additionalProperties: false,
-      },
-      {
-        properties: {
-          setRateLimit: {
-            type: 'object',
-            properties: {
-              type: {
-                type: 'string',
-                enum: FIREWALL_RATE_LIMIT_TYPES,
-                errorMessage: `The rate limit type must be one of: ${FIREWALL_RATE_LIMIT_TYPES.join(', ')}`,
-              },
-              limitBy: {
-                type: 'string',
-                enum: FIREWALL_RATE_LIMIT_BY,
-                errorMessage: `The rate limit must be applied by one of: ${FIREWALL_RATE_LIMIT_BY.join(', ')}`,
-              },
-              averageRateLimit: {
-                type: 'string',
-                errorMessage: 'The averageRateLimit must be a string',
-              },
-              maximumBurstSize: {
-                type: 'string',
-                errorMessage: 'The maximumBurstSize must be a string',
-              },
-            },
-            required: ['type', 'limitBy', 'averageRateLimit', 'maximumBurstSize'],
-            additionalProperties: false,
-            errorMessage: {
-              additionalProperties: 'No additional properties are allowed in the setRateLimit object',
-              required:
-                "All fields ('type', 'limitBy', 'averageRateLimit', 'maximumBurstSize') are required in setRateLimit",
-            },
-          },
-        },
-        required: ['setRateLimit'],
-        additionalProperties: false,
-      },
-      {
-        properties: {
-          deny: {
-            type: 'boolean',
-            const: true,
-            errorMessage: 'The deny behavior must be true',
-          },
-        },
-        required: ['deny'],
-        additionalProperties: false,
-      },
-      {
-        properties: {
-          drop: {
-            type: 'boolean',
-            const: true,
-            errorMessage: 'The drop behavior must be true',
-          },
-        },
-        required: ['drop'],
-        additionalProperties: false,
-      },
-      {
-        properties: {
-          setCustomResponse: {
-            type: 'object',
-            properties: {
-              statusCode: {
-                type: ['integer', 'string'],
-                minimum: 200,
-                maximum: 499,
-                errorMessage: 'The statusCode must be a number or string between 200 and 499',
-              },
-              contentType: {
-                type: 'string',
-                errorMessage: 'The contentType must be a string',
-              },
-              contentBody: {
-                type: 'string',
-                errorMessage: 'The contentBody must be a string',
-              },
-            },
-            required: ['statusCode', 'contentType', 'contentBody'],
-            additionalProperties: false,
-            errorMessage: {
-              additionalProperties: 'No additional properties are allowed in the setCustomResponse object',
-              required: "All fields ('statusCode', 'contentType', 'contentBody') are required in setCustomResponse",
-            },
-          },
-        },
-        required: ['setCustomResponse'],
-        additionalProperties: false,
-      },
-    ],
-    errorMessage: 'Each behavior item must contain exactly one behavior type',
-  },
-  errorMessage: 'The behaviors array must contain between 1 and 10 behavior items.',
-};
-
-const firewallFunctionsInstances = {
-  type: 'object',
-  properties: {
-    name: {
-      type: 'string',
-      minLength: 1,
-      maxLength: 100,
-      errorMessage: 'The name must be a string between 1 and 100 characters long',
-    },
-    args: {
-      type: 'object',
-      default: {},
-      errorMessage: "The 'args' field must be an object",
-    },
-    ref: {
-      type: ['string', 'number'],
-      errorMessage: "The 'ref' field must be a string or number referencing an existing Function name or ID",
-    },
-  },
-  required: ['name', 'ref'],
-  additionalProperties: false,
-  errorMessage: {
-    additionalProperties: 'No additional properties are allowed in the firewallFunctionsInstance object',
   },
 };
 
@@ -1336,161 +1168,7 @@ const azionConfigSchema = {
             },
           },
         },
-        firewall: {
-          type: 'array',
-          items: {
-            type: 'object',
-            properties: {
-              name: {
-                type: 'string',
-                errorMessage: "The firewall configuration must have a 'name' field of type string",
-              },
-              active: {
-                type: 'boolean',
-                errorMessage: "The firewall's 'active' field must be a boolean",
-              },
-              debugRules: {
-                type: 'boolean',
-                errorMessage: "The firewall's 'debugRules' field must be a boolean",
-              },
-              functions: {
-                type: 'boolean',
-                errorMessage: "The firewall's 'functions' field must be a boolean",
-              },
-              networkProtection: {
-                type: 'boolean',
-                errorMessage: "The firewall's 'networkProtection' field must be a boolean",
-              },
-              waf: {
-                type: 'boolean',
-                errorMessage: "The firewall's 'waf' field must be a boolean",
-              },
-              rules: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: {
-                    name: {
-                      type: 'string',
-                      errorMessage: "Each firewall rule must have a 'name' field of type string",
-                    },
-                    description: {
-                      type: 'string',
-                      errorMessage: "The rule's 'description' field must be a string",
-                    },
-                    active: {
-                      type: 'boolean',
-                      errorMessage: "The rule's 'active' field must be a boolean",
-                    },
-                    match: {
-                      type: 'string',
-                      errorMessage: "The rule's 'match' field must be a string containing a valid regex pattern",
-                    },
-                    variable: {
-                      anyOf: [
-                        {
-                          type: 'string',
-                          pattern: '^\\$\\{(' + [...new Set(FIREWALL_VARIABLES)].join('|') + ')\\}$',
-                          errorMessage: "The 'variable' field must be a valid variable wrapped in ${}",
-                        },
-                        {
-                          type: 'string',
-                          enum: [...FIREWALL_VARIABLES],
-                          errorMessage: "The 'variable' field must be a valid firewall variable",
-                        },
-                      ],
-                      errorMessage: "The 'variable' field must be a valid variable (with or without ${})",
-                    },
-                    behaviors: firewallRulesBehaviorsSchema,
-                    criteria: {
-                      type: 'array',
-                      minItems: 1,
-                      maxItems: 5,
-                      items: {
-                        type: 'object',
-                        properties: {
-                          conditional: {
-                            type: 'string',
-                            enum: FIREWALL_RULE_CONDITIONAL,
-                            errorMessage: `The 'conditional' field must be one of: ${FIREWALL_RULE_CONDITIONAL.join(', ')}`,
-                          },
-                          variable: {
-                            anyOf: [
-                              {
-                                type: 'string',
-                                pattern: '^\\$\\{(' + [...new Set(FIREWALL_VARIABLES)].join('|') + ')\\}$',
-                                errorMessage: "The 'variable' field must be a valid variable wrapped in ${}",
-                              },
-                              {
-                                type: 'string',
-                                enum: [...FIREWALL_VARIABLES],
-                                errorMessage: "The 'variable' field must be a valid firewall variable",
-                              },
-                            ],
-                            errorMessage: "The 'variable' field must be a valid variable (with or without ${})",
-                          },
-                          operator: {
-                            type: 'string',
-                            enum: FIREWALL_RULE_OPERATORS,
-                            errorMessage: `The 'operator' field must be one of: ${FIREWALL_RULE_OPERATORS.join(', ')}`,
-                          },
-                          argument: {
-                            type: ['string', 'number'],
-                            errorMessage: 'The argument must be a string or number',
-                          },
-                        },
-                        required: ['conditional', 'variable', 'operator', 'argument'],
-                        additionalProperties: false,
-                        errorMessage: {
-                          additionalProperties: 'No additional properties are allowed in the criteria object',
-                          required:
-                            "The 'variable', 'operator', 'argument' and 'conditional' fields are required in each criteria object",
-                        },
-                      },
-                      errorMessage: {
-                        type: 'The criteria field must be an array with at least one criteria item',
-                      },
-                    },
-                  },
-                  required: ['name', 'behaviors'],
-                  oneOf: [
-                    {
-                      required: ['match', 'variable'],
-                      not: { required: ['criteria'] },
-                      errorMessage:
-                        "When using 'match' or 'variable', both fields are required and cannot be used with 'criteria'.",
-                    },
-                    {
-                      required: ['criteria'],
-                      not: {
-                        anyOf: [{ required: ['match'] }, { required: ['variable'] }],
-                      },
-                      errorMessage: "Cannot use 'criteria' together with 'match' or 'variable'.",
-                    },
-                  ],
-                  errorMessage: {
-                    oneOf:
-                      "You must use either 'match' AND 'variable' together OR 'criteria', but not both at the same time",
-                  },
-                },
-              },
-              functionsInstances: {
-                type: 'array',
-                items: firewallFunctionsInstances,
-              },
-            },
-            required: ['name'],
-            additionalProperties: false,
-            errorMessage: {
-              type: 'Each firewall item must be an object',
-              additionalProperties: 'No additional properties are allowed in the firewall object',
-              required: "The 'name' field is required in each  firewall object",
-            },
-          },
-          errorMessage: {
-            type: "The 'firewall' field must be an array of  firewall objects",
-          },
-        },
+        firewall: firewallSchema,
         networkList: {
           type: 'array',
           items: {
