+++ packages/config/src/configProcessor/schemas/schemaManifest.ts
